extract `VmValue` from `chunk.rs` to separate `values.rs`
