+++ crates/rl-vm/src/compiler.rs
@@ -1,6 +1,7 @@
 use std::rc::Rc;
 
-use crate::chunk::{Chunk, OpCode, VmFunction, VmValue};
+use crate::chunk::{Chunk, OpCode};
+use crate::values::{VmFunction, VmValue};
 use rl_ast::{
     Ast, ExprId, nodes::ExpressionKind, statements::Statement, statements::StatementKind,
 };
