+++ crates/rl-vm/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod chunk;
 pub mod compiler;
+pub mod values;
 pub mod vm_logic;
 
-pub use chunk::{Chunk, OpCode, VmValue};
+pub use chunk::{Chunk, OpCode};
 pub use compiler::{CompileError, Compiler};
+pub use values::{VmNativeFn, VmValue};
 pub use vm_logic::{Vm, VmError};
